Update hashmap.c
This version first frees root (which is pointed to by hmap), and then hmap itself. Freeing hmap on its own is not enough (because it points to root which will leak). An alternative is:

```
sw_free(root);
sw_free(hmap);
```

But that's a little scarier.

Finally, there's also `swHashMap_free(hmap)`, which will go into `hmap` and free everything. But because infer doesn't detect this, we can rule it out.